Bot/requirements.txt
aiogram==2.19
aiohttp==3.8.1
aiosignal==1.2.0
albumentations==1.1.0
anyio==3.5.0
async-timeout==4.0.2
Cython==0.29.28
h5py==3.6.0
ipykernel==6.9.1
ipython==8.1.0
json5==0.9.6
jsonschema==4.4.0
keras==2.8.0
Keras-Preprocessing==1.1.2
Markdown==3.3.6
MarkupSafe==2.1.0
nest-asyncio==1.5.4
numpy==1.22.2
opencv-python==4.5.5.64
opencv-python-headless==4.5.5.64
pandas==1.4.1
Pillow==9.0.1
readme-renderer==34.0
tensorflow==2.8.0
testpath==0.6.0
text-unidecode==1.3
urllib3==1.26.8
utils==1.0.1